[#6]/refactor: filename을 변경합니다.
