+++ src/apis/Image/dto/image.input.ts
@@ -1,5 +1,5 @@
 import { InputType, OmitType } from '@nestjs/graphql';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
 
 @InputType()
 export class ImageInput extends OmitType(ImageEntity, ['id'], InputType) {}
