+++ src/apis/Image/images.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
-import { ImagesResolver } from '@src/apis/Image/images.resolver';
-import { ImagesService } from '@src/apis/Image/images.service';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
+import { ImagesResolver } from '@src/apis/image/images.resolver';
+import { ImagesService } from '@src/apis/image/images.service';
 
 @Module({
   imports: [
