+++ src/apis/Image/images.resolver.ts
@@ -1,6 +1,6 @@
 import { Args, Mutation, Resolver } from '@nestjs/graphql';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
-import { ImagesService } from '@src/apis/Image/images.service';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
+import { ImagesService } from '@src/apis/image/images.service';
 
 @Resolver()
 export class ImagesResolver {
