+++ src/apis/attend-list/attend-list.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AttendListResolver } from '@src/apis/attend_list/attend_list.resolver';
-import { AttendListService } from '@src/apis/attend_list/attend_list.service';
-import { AttendListEntity } from '@src/apis/attend_list/entities/attend_list.entity';
+import { AttendListResolver } from '@src/apis/attend-list/attend-list.resolver';
+import { AttendListService } from '@src/apis/attend-list/attend-list.service';
+import { AttendListEntity } from '@src/apis/attend-list/entities/attend-list.entity';
 
 @Module({
   imports: [TypeOrmModule.forFeature([AttendListEntity])],
