+++ src/apis/attend-list/attend-list.resolver.ts
@@ -1,5 +1,5 @@
 import { Resolver } from '@nestjs/graphql';
-import { AttendListService } from '@src/apis/attend_list/attend_list.service';
+import { AttendListService } from '@src/apis/attend-list/attend-list.service';
 
 @Resolver()
 export class AttendListResolver {
