+++ src/apis/auth/auth.module.ts
@@ -2,11 +2,11 @@ import { CacheModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
 import { AuthController } from '@src/apis/auth/auth.controller';
 import { AuthResolver } from '@src/apis/auth/auth.resolver';
 import { AuthService } from '@src/apis/auth/auth.service';
-import { FollowCountEntity } from '@src/apis/followCounts/followCount.entity';
+import { FollowCountEntity } from '@src/apis/follow-counts/follow-count.entity';
 import { MailService } from '@src/apis/mail/mail.service';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 import { UserService } from '@src/apis/user/user.service';
