+++ src/apis/board/dto/create-board.input.ts
