+++ src/apis/board/dto/update-board.input.ts
@@ -1,5 +1,5 @@
 import { InputType, PartialType } from '@nestjs/graphql';
-import { CreateBoardInput } from '@src/apis/board/dto/createBoard.input';
+import { CreateBoardInput } from '@src/apis/board/dto/create-board.input';
 
 @InputType()
 export class UpdateBoardInput extends PartialType(CreateBoardInput) {}
