+++ src/apis/comment/entities/comment.entity.ts
@@ -1,6 +1,6 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 import { BoardEntity } from '@src/apis/board/entities/board.entity';
-import { FictionBoardEntity } from '@src/apis/fiction_board/entities/fiction_board.entity';
+import { FictionBoardEntity } from '@src/apis/fiction-board/entities/fiction-board.entity';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 import {
   Column,
