+++ src/apis/fiction-board/dto/update-board.input.ts
@@ -1,5 +1,5 @@
 import { InputType, PartialType } from '@nestjs/graphql';
-import { CreateFictionBoardInput } from '@src/apis/fiction_board/dto/createBoard.input';
+import { CreateFictionBoardInput } from '@src/apis/fiction-board/dto/create-board.input';
 
 @InputType()
 export class UpdateFictionBoardInput extends PartialType(
