+++ src/apis/fiction-board/entities/fiction-board.entity.ts
@@ -1,5 +1,5 @@
 import { Field, Int, ObjectType } from '@nestjs/graphql';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
 import { AttendListEntity } from '@src/apis/attend_list/entities/attend_list.entity';
 import { PickEntity } from '@src/apis/pick/entities/pick.entity';
 import { TagEntity } from '@src/apis/tag/entities/tag.entity';
