+++ src/apis/fiction-board/fiction-board.service.ts
@@ -5,7 +5,7 @@ import {
   Injectable,
 } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { FictionBoardEntity } from '@src/apis/fiction_board/entities/fiction_board.entity';
+import { FictionBoardEntity } from '@src/apis/fiction-board/entities/fiction-board.entity';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 import { Like, Repository } from 'typeorm';
