+++ src/apis/follow/follow.module.ts
@@ -3,7 +3,7 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { FollowEntity } from '@src/apis/follow/entities/follow.entity';
 import { FollowResolver } from '@src/apis/follow/follow.resolver';
 import { FollowService } from '@src/apis/follow/follow.service';
-import { FollowCountEntity } from '@src/apis/followCounts/followCount.entity';
+import { FollowCountEntity } from '@src/apis/follow-counts/follow-count.entity';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 
 @Module({
