+++ src/apis/mail/mail.module.ts
@@ -1,7 +1,7 @@
 import { CacheModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
 import { MailResolver } from '@src/apis/mail/mail.resolver';
 import { MailService } from '@src/apis/mail/mail.service';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
