+++ src/apis/user/user.module.ts
@@ -1,7 +1,7 @@
 import { CacheModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { ImageEntity } from '@src/apis/Image/entities/image.entity';
+import { ImageEntity } from '@src/apis/image/entities/image.entity';
 import { BoardEntity } from '@src/apis/board/entities/board.entity';
 import { FollowCountEntity } from '@src/apis/followCounts/followCount.entity';
 import { MailService } from '@src/apis/mail/mail.service';
